Create graph.py
Here’s how you can use matplotlib to create a simple live-feed graph from a serial input. It’s quick, easy, and is a great visual.

This is some code that reads in input from a serial, and graphs that input versus a dummy “time” axis (note: this isn’t real-time, it’s incremented using code, but can be easily adapted on your end to utilize real-time if you need).